refactor(discover): polish mobile search and near me modal
