--- src/components/discover/NearMeModal.tsx
@@ -1,218 +1,0 @@
-import { lazy, Suspense, useState, type ComponentType } from "react"
-import BrutalButton from "@/components/ui/brutal-button"
-import {
-    Dialog,
-    DialogContent,
-    DialogDescription,
-    DialogHeader,
-    DialogTitle,
-} from "@/components/ui/dialog"
-import { Input } from "@/components/ui/input"
-import { useNearbyDiscoverEvents } from "@/hooks/event/useNearbyDiscoverEvents"
-import type { DiscoverLocation } from "@/types/discover"
-import {
-    formatDiscoverDate,
-    getDiscoverEventPosition,
-    getDiscoverTypeTitle,
-} from "@/components/discover/discover-utils"
-import { Crosshair, LoaderCircle, MapPinned } from "lucide-react"
-import { useNavigate } from "react-router"
-import type { MapWithGeocoderProps } from "@/components/MapWithGeocoder"
-import { testIds } from "@/testIds"
-
-type NearMeModalProps = {
-    open: boolean
-    onOpenChange: (value: boolean) => void
-    MapComponent?: ComponentType<MapWithGeocoderProps>
-}
-
-const LazyMapWithGeocoder = lazy(() => import("@/components/MapWithGeocoder"))
-
-export default function NearMeModal({
-    open,
-    onOpenChange,
-    MapComponent,
-}: NearMeModalProps) {
-    const navigate = useNavigate()
-    const ResolvedMapComponent = MapComponent ?? LazyMapWithGeocoder
-    const [radius, setRadius] = useState(3)
-    const [selectedLocation, setSelectedLocation] = useState<DiscoverLocation>()
-    const [geoLoading, setGeoLoading] = useState(false)
-    const [locationError, setLocationError] = useState<string | null>(null)
-    const {
-        data: events,
-        loading,
-        error: nearbyError,
-    } = useNearbyDiscoverEvents({
-        enabled: open,
-        query: selectedLocation
-            ? {
-                lat: selectedLocation.coordinates.lat,
-                lng: selectedLocation.coordinates.lng,
-                radius,
-            }
-            : undefined,
-    })
-    const error = locationError ?? nearbyError
-
-    function handleUseMyLocation() {
-        if (!navigator.geolocation) {
-            setLocationError("Geolocation is not supported in this browser.")
-            return
-        }
-
-        setGeoLoading(true)
-        setLocationError(null)
-
-        navigator.geolocation.getCurrentPosition(
-            (position) => {
-                setSelectedLocation({
-                    value: "Current location",
-                    coordinates: {
-                        lat: position.coords.latitude,
-                        lng: position.coords.longitude,
-                    },
-                })
-                setGeoLoading(false)
-            },
-            () => {
-                setGeoLoading(false)
-                setLocationError("We couldn't access your location. Search for a place instead.")
-            },
-        )
-    }
-
-    const markers = events.map((event) => ({
-        id: event._id,
-        position: getDiscoverEventPosition(event),
-        title: event.title,
-        description: event.fullAddress,
-        href: `/event/${event._id}`,
-    }))
-
-    function handleOpenEvent(eventId: string) {
-        onOpenChange(false)
-        navigate(`/event/${eventId}`)
-    }
-
-    return (
-        <Dialog open={open} onOpenChange={onOpenChange}>
-            <DialogContent className="left-0 top-0 h-[100svh] w-screen max-h-[100svh] max-w-none translate-x-0 translate-y-0 gap-0 overflow-hidden rounded-none border-0 bg-card p-0 shadow-none [&>button]:right-4 [&>button]:top-4 sm:left-[50%] sm:top-[50%] sm:h-auto sm:w-[94vw] sm:max-h-[92vh] sm:max-w-6xl sm:-translate-x-1/2 sm:-translate-y-1/2 sm:border-brutal sm:shadow-brutal-xl sm:[&>button]:right-5 sm:[&>button]:top-5">
-                <DialogHeader className="sr-only">
-                    <DialogTitle>Near me</DialogTitle>
-                    <DialogDescription>Explore events near your location on the map.</DialogDescription>
-                </DialogHeader>
-
-                <div className="flex items-center gap-2 border-b-2 border-border bg-cream px-4 py-2.5 pr-14 sm:px-5 sm:py-5 sm:pr-16">
-                    <span className="h-3 w-3 rounded-full border border-border bg-[#FF6B6B]" />
-                    <span className="h-3 w-3 rounded-full border border-border bg-yellow" />
-                    <span className="h-3 w-3 rounded-full border border-border bg-mint" />
-                </div>
-
-                <div className="flex min-h-0 flex-1 flex-col lg:grid lg:grid-cols-[22rem_minmax(0,1fr)]">
-                    <div className="order-2 min-h-0 flex-1 space-y-5 overflow-y-auto border-t-2 border-border bg-cream p-2.5 lg:order-1 lg:border-r-2 lg:border-t-0 lg:p-5">
-                        <div>
-                            <p className="mb-2 font-mono text-[11px] font-bold uppercase tracking-[0.18em] text-violet">
-                                Nearby events
-                            </p>
-                            <h3 className="font-display text-[2.2rem] font-extrabold leading-[0.95] tracking-[-0.06em] lg:text-4xl">
-                                Find what is happening around you.
-                            </h3>
-                            <p className="mt-3 text-sm leading-6 text-foreground/70">
-                                Use your current location or search for another place to preview events on the map.
-                            </p>
-                        </div>
-
-                        <BrutalButton
-                            tone="mint"
-                            data-testid={testIds.discover.nearMeUseLocationButton}
-                            className="w-full gap-2"
-                            onClick={handleUseMyLocation}
-                            disabled={geoLoading}
-                        >
-                            {geoLoading ? <LoaderCircle className="size-4 animate-spin" /> : <Crosshair className="size-4" />}
-                            Use my location
-                        </BrutalButton>
-
-                        <div className="space-y-2">
-                            <label className="font-mono text-[11px] font-bold uppercase tracking-[0.18em] text-foreground/65">
-                                Radius (1-15)
-                            </label>
-                            <Input
-                                type="number"
-                                min={1}
-                                max={15}
-                                value={radius}
-                                onChange={(event) => setRadius(Math.min(15, Math.max(1, Number(event.target.value) || 1)))}
-                                className="h-11 rounded-none border-brutal border-border bg-card"
-                            />
-                        </div>
-
-                        {selectedLocation ? (
-                            <div className="border-2 border-border bg-card p-4 shadow-brutal-sm">
-                                <p className="font-mono text-[11px] font-bold uppercase tracking-[0.18em] text-foreground/60">
-                                    Searching around
-                                </p>
-                                <p className="mt-2 text-sm font-semibold">{selectedLocation.value}</p>
-                            </div>
-                        ) : null}
-
-                        {error ? (
-                            <div className="border-2 border-border bg-coral px-4 py-3 text-sm font-semibold text-white">
-                                {error}
-                            </div>
-                        ) : null}
-
-                        <div className="space-y-3">
-                            <div className="flex items-center justify-between">
-                                <p className="font-heading text-lg font-bold">Nearby results</p>
-                                {loading ? <LoaderCircle className="size-4 animate-spin" /> : null}
-                            </div>
-                            <div className="max-h-56 space-y-3 overflow-y-auto pr-1 lg:max-h-64">
-                                {!loading && events.length === 0 ? (
-                                    <div className="border-2 border-dashed border-border bg-card px-4 py-5 text-sm text-foreground/70">
-                                        Pick a location to load nearby events.
-                                    </div>
-                                ) : null}
-                                {events.map((event) => (
-                                    <button
-                                        key={event._id}
-                                        type="button"
-                                        data-testid={testIds.discover.nearMeResultButton(event._id)}
-                                        onClick={() => handleOpenEvent(event._id)}
-                                        className="w-full border-2 border-border bg-card p-4 text-left shadow-brutal-sm transition-transform hover:-translate-y-1 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-violet"
-                                    >
-                                        <div className="flex items-start justify-between gap-3">
-                                            <div>
-                                                <p className="font-heading text-base font-bold">{event.title}</p>
-                                                <p className="mt-1 text-xs uppercase tracking-[0.14em] text-foreground/55">
-                                                    {getDiscoverTypeTitle(event)}
-                                                </p>
-                                            </div>
-                                            <MapPinned className="size-4 shrink-0" />
-                                        </div>
-                                        <p className="mt-2 text-sm text-foreground/75">{event.fullAddress}</p>
-                                        <p className="mt-2 text-sm font-semibold">{formatDiscoverDate(event.date)}</p>
-                                    </button>
-                                ))}
-                            </div>
-                        </div>
-                    </div>
-
-                    <div className="order-1 shrink-0 bg-violet-light p-1 sm:p-3 md:p-5 lg:order-2">
-                        <div className="h-[19svh] min-h-[10.5rem] overflow-hidden border-2 border-border bg-card shadow-brutal sm:h-[28rem] lg:h-[calc(92vh-9rem)] lg:min-h-[30rem]">
-                            <Suspense fallback={<div className="h-full w-full bg-card" />}>
-                                <ResolvedMapComponent
-                                    value={selectedLocation}
-                                    onChange={setSelectedLocation}
-                                    markers={markers}
-                                    height="100%"
-                                />
-                            </Suspense>
-                        </div>
-                    </div>
-                </div>
-            </DialogContent>
-        </Dialog>
-    )
-}
