+++ src/pages/Events.spec.tsx
@@ -3,7 +3,7 @@ import test, { type TestContext } from "node:test"
 import { fireEvent, render, waitFor, within } from "@testing-library/react"
 import { createMemoryRouter, RouterProvider } from "react-router"
 import DiscoverFilterBar from "@/components/discover/DiscoverFilterBar"
-import NearMeModal from "@/components/discover/NearMeModal"
+import NearMeModal from "@/components/discover/nearMeModal/NearMeModal"
 import Events from "@/pages/Events"
 import { api } from "@/lib/axios"
 import { routes as eventRoutes } from "@/routes/events"
