+++ src/pages/Events.tsx
@@ -2,7 +2,7 @@ import { useDeferredValue, useState } from "react"
 import DiscoverEventList from "@/components/discover/DiscoverEventList"
 import DiscoverFilterBar from "@/components/discover/DiscoverFilterBar"
 import DiscoverResultsHeader from "@/components/discover/DiscoverResultsHeader"
-import NearMeModal from "@/components/discover/NearMeModal"
+import NearMeModal from "@/components/discover/nearMeModal/NearMeModal"
 import { useDiscoverCategories, useDiscoverEventTypes } from "@/hooks/event/useDiscoverEventMetadata"
 import { useDiscoverEvents } from "@/hooks/event/useDiscoverEvents"
 import { useDiscoverFilters } from "@/hooks/event/useDiscoverFilters"
